# Remove Unnecessary INNER JOIN in Many To Many assoc_query