--- a/lib/ecto/association.ex
+++ b/lib/ecto/association.ex
@@ -134,7 +134,7 @@ defmodule Ecto.Association do
   Returns information used by the preloader.
   """
   @callback preload_info(t) ::
-              {:assoc, t, {integer, atom}} | {:through, t, [atom]}
+              {:assoc, t, {integer, atom}} | {:through, t, [atom]} | {:assoc, t, {integer, atom, Ecto.Type.t()}}
 
   @doc """
   Performs the repository change on the association.
@@ -1139,14 +1139,16 @@ defmodule Ecto.Association.ManyToMany do
 
   @impl true
   def assoc_query(assoc, query, values) do
-    %{queryable: queryable, join_through: join_through, join_keys: join_keys} = assoc
-    [{join_owner_key, _}, {join_related_key, related_key}] = join_keys
+    %{queryable: queryable, join_through: join_through, join_keys: join_keys, owner: owner} = assoc
+    [{join_owner_key, owner_key}, {join_related_key, related_key}] = join_keys
+
+    owner_key_type = owner.__schema__(:type, owner_key)
 
     # We only need to join in the "join table". Preload and Ecto.assoc expressions can then filter
     # by &1.join_owner_key in ^... to filter down to the associated entries in the related table.
     from(q in (query || queryable),
       join: j in ^join_through, on: field(q, ^related_key) == field(j, ^join_related_key),
-      where: field(j, ^join_owner_key) in ^values
+      where: field(j, ^join_owner_key) in type(^values, {:in, ^owner_key_type})
     )
     |> Ecto.Association.combine_assoc_query(assoc.where)
     |> Ecto.Association.combine_joins_query(assoc.join_where, 1)
@@ -1160,10 +1162,12 @@ defmodule Ecto.Association.ManyToMany do
   end
 
   @impl true
-  def preload_info(%{join_keys: [{join_owner_key, _}, {_, _}]} = refl) do
+  def preload_info(%{join_keys: [{join_owner_key, owner_key}, {_, _}], owner: owner} = refl) do
+    owner_key_type = owner.__schema__(:type, owner_key)
+
     # When preloading use the last bound table (which is the join table) and the join_owner_key
     # to filter out related entities to the owner structs we're preloading with.
-    {:assoc, refl, {-1, join_owner_key}}
+    {:assoc, refl, {-1, join_owner_key, owner_key_type}}
   end
 
   @impl true