@josevalim 2020-11-14
Thanks!

Did you run ecto_sql suite as specified in the README? You can set ECTO_PATH pointing to your checkout.