@GabrielAlacchi 2020-11-14
Thanks @josevalim this one took some head scratching. I am getting an integration testing error let me investigate it. Have you tried making a github workflow to run these tests? I can take a crack at that later if you'd like.

```
1) test many_to_many changeset assoc with self-referential binary_id (Ecto.Integration.AssocTest)
     /home/gabriel/oss/ecto/integration_test/cases/assoc.exs:470
     ** (DBConnection.EncodeError) Postgrex expected a binary of 16 bytes, got "6379f124-0984-410d-b127-81d7e8fbdbc2". Please make sure the value you are passing matches the definition in your table or in your query or convert the value accordingly.
     code: custom = Custom |> TestRepo.get!(custom.bid) |> TestRepo.preload(:customs)
     stacktrace:
       (postgrex 0.15.5) lib/postgrex/type_module.ex:897: Postgrex.DefaultTypes.encode_list/3
       (postgrex 0.15.5) lib/postgrex/extensions/array.ex:82: Postgrex.Extensions.Array.encode/4
       (postgrex 0.15.5) lib/postgrex/extensions/array.ex:52: Postgrex.Extensions.Array.encode/3
       (postgrex 0.15.5) lib/postgrex/type_module.ex:897: Postgrex.DefaultTypes.encode_params/3
       (postgrex 0.15.5) lib/postgrex/query.ex:75: DBConnection.Query.Postgrex.Query.encode/3
       (db_connection 2.2.2) lib/db_connection.ex:1148: DBConnection.encode/5
       (db_connection 2.2.2) lib/db_connection.ex:1246: DBConnection.run_prepare_execute/5
       (db_connection 2.2.2) lib/db_connection.ex:1342: DBConnection.run/6
       (db_connection 2.2.2) lib/db_connection.ex:539: DBConnection.parsed_prepare_execute/5
       (db_connection 2.2.2) lib/db_connection.ex:532: DBConnection.prepare_execute/4
       (ecto_sql 3.6.0-dev) lib/ecto/adapters/sql.ex:682: Ecto.Adapters.SQL.execute!/4
       (ecto_sql 3.6.0-dev) lib/ecto/adapters/sql.ex:674: Ecto.Adapters.SQL.execute/5
       (ecto 3.5.4) lib/ecto/repo/queryable.ex:229: Ecto.Repo.Queryable.execute/4
       (ecto 3.5.4) lib/ecto/repo/queryable.ex:17: Ecto.Repo.Queryable.all/3
       (elixir 1.11.1) lib/enum.ex:1399: Enum."-map/2-lists^map/1-0-"/2
       /home/gabriel/oss/ecto/integration_test/cases/assoc.exs:474: (test)
```