@josevalim 2020-11-14
Is that the only failure you get? So the reason why it happens is that, because we are not going through the scheme, we are no long dumping the type when join_through is a binary. We may need to write the query like this: where: field in type(^ids, {:in, type_of_pk}) or similar. Not sure how easy it is, I can take a look later.