@GabrielAlacchi 2020-11-14
Trying this out, the query passes type checking but the preload is failing to load the associated record in this case.

So the `assoc_query` now looks like

```
    from(q in (query || queryable),
      join: j in ^join_through, on: field(q, ^related_key) == field(j, ^join_related_key),
      where: field(j, ^join_owner_key) in type(^values, {:in, ^owner_key_type})
    )
    |> Ecto.Association.combine_assoc_query(assoc.where)
    |> Ecto.Association.combine_joins_query(assoc.join_where, 1)
```

````
1) test many_to_many changeset assoc with self-referential binary_id (Ecto.Integration.AssocTest)
     /home/gabriel/oss/ecto/integration_test/cases/assoc.exs:470
     match (=) failed
     code:  assert [_] = custom.customs
     left:  [_]
     right: []
     stacktrace:
       /home/gabriel/oss/ecto/integration_test/cases/assoc.exs:476: (test)
```