@josevalim 2020-11-14
You can enable logging in the test_helper.exs file of each adapter. Just comment the first line.