@GabrielAlacchi 2020-11-15
@josevalim My latest commit seems to have solved the problem with regards to the type checking with the binary id, although I'm not sure if it's the best solution. My approach was to use `type(^values, {:in, owner_key_type})` as you suggested. I also needed to tell the preloading query that the `select` of the related key has the same type as `owner_key_type`, and so I modified the definition for `related_key_to_field` to specifically add a `type` call to the AST to tell it that the related field has that type.

There remains some integration testing errors for `myxql` but these don't seem to be related to this PR? I ran against master and these still showed up.

```
  1) test decimal type cast (Ecto.Integration.TypeTest)
     /home/gabriel/oss/ecto/integration_test/cases/type.exs:424
     Assertion with == failed
     code:  assert TestRepo.all(from(p in Post, select: type(2 + ^"2", p.cost))) == [Decimal.new("4")]
     left:  [#Decimal<4.000000000000000000000000000000>]
     right: [#Decimal<4>]
     stacktrace:
       /home/gabriel/oss/ecto/integration_test/cases/type.exs:428: (test)

..............................................................................................................................................................................................................................................

  2) test bit as boolean (Ecto.Integration.MyXQLTypeTest)
     integration_test/myxql/myxql_type_test.exs:23
     ** (MyXQL.Error) (1210) (ER_WRONG_ARGUMENTS) Incorrect arguments to mysqld_stmt_execute
     code: TestRepo.insert_all("bits", [[bit: <<1::1>>], [bit: <<0::1>>]])
     stacktrace:
       (ecto_sql 3.6.0-dev) lib/ecto/adapters/sql.ex:741: Ecto.Adapters.SQL.raise_sql_call_error/1
       (ecto_sql 3.6.0-dev) lib/ecto/adapters/sql.ex:650: Ecto.Adapters.SQL.insert_all/8
       (ecto 3.5.4) lib/ecto/repo/schema.ex:54: Ecto.Repo.Schema.do_insert_all/6
       integration_test/myxql/myxql_type_test.exs:24: (test)



  3) test bit (Ecto.Integration.MyXQLTypeTest)
     integration_test/myxql/myxql_type_test.exs:14
     ** (MyXQL.Error) (1210) (ER_WRONG_ARGUMENTS) Incorrect arguments to mysqld_stmt_execute
     code: TestRepo.insert_all("bits", [[bit: <<1::1>>], [bit: <<0::1>>]])
     stacktrace:
       (ecto_sql 3.6.0-dev) lib/ecto/adapters/sql.ex:741: Ecto.Adapters.SQL.raise_sql_call_error/1
       (ecto_sql 3.6.0-dev) lib/ecto/adapters/sql.ex:650: Ecto.Adapters.SQL.insert_all/8 
       (ecto 3.5.4) lib/ecto/repo/schema.ex:54: Ecto.Repo.Schema.do_insert_all/6
       integration_test/myxql/myxql_type_test.exs:15: (test)
```