@tcrossland 2021-04-06
Just to let you know, this change breaks certain expression queries. For example, the following no longer works:

```elixir
parent
|> Ecto.assoc(:children)
|> join(:inner, [child], grandchild in assoc(child, :children))
|> where([_parent, _child, rel, _grandchild], rel.type_id == 0)
|> Repo.all()
```

... where the parent schema has a `many_to_many` relation with itself via a `join_through` schema.